requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-rest-framework==0.1.0
djangorestframework==3.11.0
get==2019.4.13
idna==2.9
pkg-resources==0.0.0
post==2019.4.13
public==2019.4.13
python-dotenv==0.13.0
pytz==2020.1
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
